Fixed and issue when sending a packet to the channel. The load was not set correctly

+++ Aux.h
@@ -5,7 +5,7 @@ struct Packet
 {
 	int source;
 	int destination;
-	int L;
+	long int L;
 	long int seq;
 	double tx_time; //time at which a packet is transmitted
 	double queuing_time; //time at which the packet is deposited at end of the queue
